Revert "seperated react files"
This reverts commit 56c2f384a1c06313176e96b03cf3f6cfa2b412c9.
