+++ .gitignore
@@ -21,6 +21,3 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-
-/k8s/dash
-skaffold.yaml
