--- client/src/components/MovieCard/MovieCard.jsx
@@ -1,14 +1,0 @@
-  const MovieCard = ({id, title, poster}) => {
-    
-    return (
-      
-              <div key={id} className="movie__card">
-                  <p className="movie__title">{title}</p>
-                  <img className="movie__poster" src={poster} alt="poster" />
-              </div> 
-      
-      )
-    }
-
-
-export default MovieCard;
