+++ k8s/main/dash-clusterrole.yaml
