--- skaffold.yaml
@@ -1,22 +1,0 @@
-apiVersion: skaffold/v2beta12
-kind: Config
-deploy:
-  kubectl:
-    manifests:
-      - ./k8s/main/client-deployment.yaml
-build:
-  local:
-    push: false
-  artifacts:
-    - image: mustafansrv/multi-client
-      context: client
-      docker:
-        dockerfile: Dockerfile.dev
-      sync:
-        manual:
-          - src: "src/**/*.js"
-            dest: .
-          - src: "src/**/*.css"
-            dest: .
-          - src: "src/**/*.html"
-            dest: .
